Calculate distance spacing from data, not metadata, for FebusTools.FebusData
In constructing a `DASArray` from a `FebusTools.FebusData` object, we
previously used the file metadata to find the distance (channel)
spacing.  However, this may not be correct, because the file metadata
is copied unaltered from the original input file, whereas `t.distances`
could be different if we have downsampled the data in space upon
reading.