Makie: Add an extension for plotting with Makie
Firstly, add the functions `plot_heatmap` and `plot_heatmap!` to plot
`DASArray`s and `FourierDASArray`s as heatmaps.  At the same time,
add methods to `Makie.heatmap[!]` to call these functions for the
appropriate arguments.

Secondly, add methods to `Seis.plot_section` to plot trace arrays as
record sections and automatically pick the most appropriate values on
the y-axis at which to plot traces.

Notably, there are no tests for this plotting functionality; perhaps it
should be added in a separate workflow run or be guarded by an
environment variable so that tests can be run without plotting for the
most part.